Commit initial code for char and updated code for int

+++ vecadd/repeat_calc_int/time/pp_dynamic_access_offchip_memory_int_vecadd_repeat_kernel.cu
@@ -1,6 +1,6 @@
 #include "pp_dynamic_access_offchip_memory_int_vecadd_repeat.h"
 
-__global__ void AddVectors(const float* A, const float* B, float* C, int N)
+__global__ void AddVectors(const int* A, const int* B, int* C, int N)
 {
   int blockStartIndex  = blockIdx.x * blockDim.x * N;
   int threadStartIndex = blockStartIndex + threadIdx.x;
